+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Do not resave already-saved session at end of request
   * deps: cookie-signature@1.0.5
   * deps: debug@~2.0.0